Did the following things:
Add bumper to video text attribute field
Going from edit hobby to new hobby should clear the form
Page lags when trying to load the video. Maybe show the image thumb
until people click it.
Videos are not changing when the hobby changes
Saving a new hobby should change the url to be the edit page of the new hobby